# src/main/java/com/sprint/mission/discodeit/controller/BinaryContentController.java
package com.sprint.mission.discodeit.controller;

import com.sprint.mission.discodeit.controller.api.BinaryContentApi;
import com.sprint.mission.discodeit.dto.data.BinaryContentDTO;
import com.sprint.mission.discodeit.service.BinaryContentService;
import com.sprint.mission.discodeit.storage.BinaryContentStorage;
import java.util.List;
import java.util.UUID;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;
import lombok.extern.slf4j.Slf4j;

@Slf4j
@RequiredArgsConstructor
@RestController
@RequestMapping("/api/binaryContents")
public class BinaryContentController implements BinaryContentApi {

  private final BinaryContentService binaryContentService;
  private final BinaryContentStorage binaryContentStorage;

  @GetMapping(path = "{binaryContentId}")
  public ResponseEntity<BinaryContentDTO> find(
      @PathVariable("binaryContentId") UUID binaryContentId) {
    log.info("바이너리 컨텐츠 조회 요청: id={}", binaryContentId);
    BinaryContentDTO binaryContent = binaryContentService.find(binaryContentId);
    log.debug("바이너리 컨텐츠 조회 응답: {}", binaryContent);
    return ResponseEntity
        .status(HttpStatus.OK)
        .body(binaryContent);
  }

  @GetMapping
  public ResponseEntity<List<BinaryContentDTO>> findAllByIdIn(
      @RequestParam("binaryContentIds") List<UUID> binaryContentIds) {
    log.info("바이너리 컨텐츠 목록 조회 요청: ids={}", binaryContentIds);
    List<BinaryContentDTO> binaryContents = binaryContentService.findAllByIdIn(binaryContentIds);
    log.debug("바이너리 컨텐츠 목록 조회 응답: count={}", binaryContents.size());
    return ResponseEntity
        .status(HttpStatus.OK)
        .body(binaryContents);
  }

  @GetMapping(path = "{binaryContentId}/download")
  public ResponseEntity<?> download(
      @PathVariable("binaryContentId") UUID binaryContentId) {
    log.info("바이너리 컨텐츠 다운로드 요청: id={}", binaryContentId);
    BinaryContentDTO binaryContentDTO = binaryContentService.find(binaryContentId);
    ResponseEntity<?> response = binaryContentStorage.download(binaryContentDTO);
    log.debug("바이너리 컨텐츠 다운로드 응답: contentType={}, contentLength={}", 
        response.getHeaders().getContentType(), response.getHeaders().getContentLength());
    return response;
  }
}