# src/main/java/com/sprint/mission/discodeit/dto/data/ChannelDTO.java
package com.sprint.mission.discodeit.dto.data;

import com.sprint.mission.discodeit.entity.ChannelType;
import java.time.Instant;
import java.util.List;
import java.util.UUID;

public record ChannelDTO(
    UUID id,
    ChannelType type,
    String name,
    String description,
    List<UserDTO> participants,
    Instant lastMessageAt
) {

}